Added scale feature to blocks
