+++ python/examples/stage_test/settings.json
@@ -1,15 +1,16 @@
 {   
     "boundary_conditions":{
-        "P0_Inlet": 60,
-        "T0_Inlet": 524.87,
-        "Ps_outlet": 8.986,
-        "bl_height": 0.1366,
+        "P0_Inlet": 5,
+        "T0_Inlet": 700,
+        "Ps_outlet": 1,
+        "bl_height": 0.001,
         "Mach_Inlet":0.10,
-        "alpha":40,
+        "alpha":0,
         "refLen":0.0254,
         "Tu_intensity_init": 0.03,
         "Tu_length_scale_init": 0.03,
-        "NBlades":55
+        "NBlades":[55,60],
+        "RotorRPM":7500
     },
     "job_settings": [
         {
